tegra-helper-scripts-native: Add initrd flashing support
* tegraXXX-flash-helper scripts updated to support RCM booting,
  external device handling, and other features needed to support
  flashing via a Linux-based device-side script
* nvflashxmlparse updated for working with separate flash layouts
  for external devices and modifying flash layout files to support
  the initrd-based flashing process
* make-sdcard updated for writing to storage devices exported
  by the Jetson via USB mass-storage gadget
* find-jetson-usb script added for discovering Jetsons in
  recovery mode
* rewrite-tegraflash-args added to update the normal flashing command
  performs an RCM boot.  Needed so we can RCM boot when using bootloader
  binaries that were signed during the build
* brcid-to-uid added for extracting the ECID from the initial bootrom
  communication with a Jetson, and transforming it into the ECID exposed
  by the tegra fuse driver in the Linux kernel, so we can use it as a
  unique identifier when locating USB storage exports from the device

+++ recipes-bsp/tegra-binaries/tegra-helper-scripts/tegra234-flash-helper.sh
@@ -1,15 +1,19 @@
 #!/bin/bash
 bup_blob=0
+rcm_boot=0
 keyfile=
 sbk_keyfile=
 user_keyfile=
 spi_only=
+external_device=0
 sdcard=
 no_flash=0
+to_sign=0
 flash_cmd=
 imgfile=
 dataimg=
 inst_args=""
+extdevargs=
 blocksize=4096
 
 # These functions are used in odmsign.func but do not
@@ -24,7 +28,7 @@ cp2local() {
     :
 }
 
-ARGS=$(getopt -n $(basename "$0") -l "bup,no-flash,sdcard,spi-only,datafile:,usb-instance:,user_key:" -o "u:v:s:b:B:yc:" -- "$@")
+ARGS=$(getopt -n $(basename "$0") -l "bup,no-flash,sign,sdcard,spi-only,boot-only,external-device,rcm-boot,datafile:,usb-instance:,user_key:" -o "u:v:s:b:B:yc:" -- "$@")
 if [ $? -ne 0 ]; then
     echo "Error parsing options" >&2
     exit 1
@@ -43,12 +47,25 @@ while true; do
         no_flash=1
         shift
         ;;
+    --sign)
+        to_sign=1
+        shift
+        ;;
     --sdcard)
         sdcard=yes
         shift
         ;;
-    --spi-only)
+    --spi-only|--boot-only)
         spi_only=yes
+        shift
+        ;;
+    --rcm-boot)
+        rcm_boot=1
+        shift
+        ;;
+    --external-device)
+        external_device=1
+	extdevargs="--external_device"
         shift
         ;;
     --datafile)
@@ -127,14 +144,27 @@ if [ -z "$FLASHVARS" ]; then
     exit 1
 fi
 
+have_odmsign_func=0
+[ ! -e "$here/odmsign.func" ] || have_odmsign_func=1
+if [ -n "$keyfile" -o -n "$sbk_keyfile" -o -n "$user_keyfile" ] && [ $have_odmsign_func -eq 0 ]; then
+    echo "ERR: missing odmsign.func from secureboot package, signing not supported" >&2
+    exit 1
+fi
+
+if [ $rcm_boot -eq 1 ]; then
+    OVERLAY_DTB_FILE=$(echo "$OVERLAY_DTB_FILE" | sed -e's!BootOrder[^.]*\.dtbo\(,\|$\)!!')
+fi
+
 # Temp file for storing cvm.bin in, if we need to query the board for its
 # attributes
 cvm_bin=$(mktemp cvm.bin.XXXXX)
 
 skipuid=""
 bootauth=""
+BR_CID=
 if [ -z "$CHIPREV" -o -z "$fuselevel" ]; then
-    chipid=$($here/tegrarcm_v2 --new_session --chip 0x23 ${inst_args} --uid | grep BR_CID | cut -d' ' -f2)
+    BR_CID=$($here/tegrarcm_v2 --new_session --chip 0x23 ${inst_args} --uid | grep BR_CID | cut -d' ' -f2)
+    chipid="$BR_CID"
     if [ -z "$chipid" ]; then
         echo "ERR: could not retrieve chip ID" >&2
         exit 1
@@ -190,15 +220,18 @@ if [ -z "$CHIPREV" -o -z "$fuselevel" ]; then
             fi
             ;;
     esac
+else
+    skipuid="--skipuid"
 fi
 
+have_boardinfo=
 if [ -z "$FAB" -o -z "$BOARDID" ]; then
     if [ "$fuselevel" = "fuselevel_production" ]; then
         sed -i "s/preprod_dev_sign = <1>/preprod_dev_sign = <0>/" "${EMC_FUSE_DEV_PARAMS}";
     fi
     keyargs=
-    [ -z "$keyfile" ] || keyargs="$keyargs --key \"$keyfile\""
-    [ -z "$sbk_keyfile" ] || keyargs="$keyargs --encrypt_key \"$sbk_keyfile\""
+    [ -z "$keyfile" ] || keyargs="$keyargs --key $keyfile"
+    [ -z "$sbk_keyfile" ] || keyargs="$keyargs --encrypt_key $sbk_keyfile"
     rm -f rcm_state
     if ! python3 $flashappname ${inst_args} --chip 0x23 $skipuid $keyargs \
          --applet mb1_t234_prod.bin \
@@ -221,6 +254,7 @@ if [ -z "$FAB" -o -z "$BOARDID" ]; then
     # ramcode_id=$($here/chkbdinfo -R chip_info.bin_bak)
     # -XXX
     skipuid=""
+    have_boardinfo="yes"
 fi
 
 if [ -z "$CHIP_SKU" ]; then
@@ -242,15 +276,18 @@ else
 fi
 if [ -n "$BOARDSKU" ]; then
     board_sku="$BOARDSKU"
-else
+elif [ -n "$have_boardinfo" ]; then
     board_sku=`$here/chkbdinfo -k ${cvm_bin} | tr -d '[:space:]' | tr [a-z] [A-Z]`
     BOARDSKU="$board_sku"
 fi
-if [ "${BOARDREV+isset}" = "isset" ]; then
+if [ -n "$BOARDREV" ]; then
     board_revision="$BOARDREV"
-else
+elif [ -n "$have_boardinfo" ]; then
     board_revision=`$here/chkbdinfo -r ${cvm_bin} | tr -d '[:space:]' | tr [a-z] [A-Z]`
     BOARDREV="$board_revision"
+fi
+if [ -z "$serial_number" -a -n "$have_boardinfo" ]; then
+    serial_number=$($here/chkbdinfo -a ${cvm_bin} | tr -d '[:space:]')
 fi
 
 [ -f ${cvm_bin} ] && rm -f ${cvm_bin}
@@ -258,7 +295,26 @@ fi
 [ -n "$BOARDID" ] || BOARDID=3701
 [ -n "$FAB" ] || FAB=TS4
 [ -n "$fuselevel" ] || fuselevel=fuselevel_production
-[ -n "${BOOTDEV}" ] || BOOTDEV="mmcblk0p1"
+[ -n "$BOOTDEV" ] || BOOTDEV="mmcblk0p1"
+
+rm -f boardvars.sh
+cat >boardvars.sh <<EOF
+BOARDID="$BOARDID"
+FAB="$FAB"
+BOARDSKU="$BOARDSKU"
+BOARDREV="$BOARDREV"
+CHIPREV="$CHIPREV"
+fuselevel="$fuselevel"
+EOF
+if [ -n "$serial_number" ]; then
+    echo "serial_number=$serial_number" >>boardvars.sh
+fi
+if [ -n "$usb_instance" ]; then
+    echo "usb_instance=$usb_instance" >>boardvars.sh
+fi
+if [ -n "$BR_CID" ]; then
+    echo "BR_CID=\"$BR_CID\"" >>boardvars.sh
+fi
 
 if [ "$BOARDID" = "3701" -a "$FAB" = "301" ]; then
     RAMCODE=0
@@ -314,7 +370,7 @@ printf "0x%x\n" $(( (BSP_BRANCH<<16) | (BSP_MAJOR<<8) | BSP_MINOR )) >>${MACHINE
 bytes=$(wc -c ${MACHINE}_bootblob_ver.txt | cut -d' ' -f1)
 cksum=$(python3 -c "import zlib; print(\"%X\" % (zlib.crc32(open(\"${MACHINE}_bootblob_ver.txt\", \"rb\").read()) & 0xFFFFFFFF))")
 echo "BYTES:$bytes CRC32:$cksum" >>${MACHINE}_bootblob_ver.txt
-if [ -z "$sdcard" ]; then
+if [ -z "$sdcard" -a $external_device -eq 0 ]; then
     appfile=$(basename "$imgfile").img
     if [ -n "$dataimg" ]; then
     datafile=$(basename "$dataimg").img
@@ -324,16 +380,18 @@ else
     datafile="$dataimg"
 fi
 appfile_sed=
-if [ $bup_blob -ne 0 ]; then
+if [ $bup_blob -ne 0 -o $rcm_boot -ne 0 ]; then
     kernfile="${kernfile:-boot.img}"
     appfile_sed="-e/APPFILE/d -e/DATAFILE/d"
-elif [ $no_flash -eq 0 -a -z "$sdcard" ]; then
+elif [ $no_flash -eq 0 -a -z "$sdcard" -a $external_device -eq 0 ]; then
     appfile_sed="-es,APPFILE_b,$appfile, -es,APPFILE,$appfile, -es,DATAFILE,$datafile,"
+elif [ $no_flash -ne 0 ]; then
+    touch APPFILE APPFILE_b DATAFILE
 else
     pre_sdcard_sed="-es,APPFILE_b,$appfile, -es,APPFILE,$appfile,"
     if [ -n "$datafile" ]; then
-    pre_sdcard_sed="$pre_sdcard_sed -es,DATAFILE,$datafile,"
-    touch DATAFILE
+	pre_sdcard_sed="$pre_sdcard_sed -es,DATAFILE,$datafile,"
+	touch DATAFILE
     fi
     touch APPFILE APPFILE_b
 fi
@@ -343,12 +401,14 @@ kernel_dtbfile="kernel_$dtb_file_basename"
 rm -f "$kernel_dtbfile"
 cp "$dtb_file" "$kernel_dtbfile"
 
-if [ "$spi_only" = "yes" ]; then
+if [ "$spi_only" = "yes" -o $external_device -eq 1 ]; then
     if [ ! -e "$here/nvflashxmlparse" ]; then
-    echo "ERR: missing nvflashxmlparse script" >&2
-    exit 1
+        echo "ERR: missing nvflashxmlparse script" >&2
+        exit 1
     fi
-    "$here/nvflashxmlparse" --extract -t spi -o flash.xml.tmp "$flash_in" || exit 1
+fi
+if [ "$spi_only" = "yes" ]; then
+    "$here/nvflashxmlparse" --extract -t boot -o flash.xml.tmp "$flash_in" || exit 1
 else
     cp "$flash_in" flash.xml.tmp
 fi
@@ -370,8 +430,12 @@ sce_fw camera-rtcpu-sce.img; \
 rce_fw camera-rtcpu-t234-rce.img; \
 ape_fw adsp-fw.bin; \
 spe_fw spe_t234.bin; \
-tlk tos-optee_t234.img; \
+tos tos-optee_t234.img; \
 eks eks.img"
+
+if [ $rcm_boot -ne 0 ]; then
+    BINSARGS="$BINSARGS; kernel $kernfile; kernel_dtb $kernel_dtbfile"
+fi
 
 custinfo_args=
 if [ -f "$custinfo_out" ]; then
@@ -396,11 +460,13 @@ bctargs="$UPHY_CONFIG $MINRATCHET_CONFIG \
          --cpubl uefi_jetson.bin \
          --overlay_dtb $OVERLAY_DTB_FILE $custinfo_args"
 
-if [ $bup_blob -ne 0 -o "$sdcard" = "yes" ]; then
+if [ $bup_blob -ne 0 -o $to_sign -ne 0 -o "$sdcard" = "yes" -o $external_device -eq 1 ]; then
     tfcmd=sign
     skipuid="--skipuid"
+elif [ $rcm_boot -ne 0 ]; then
+    tfcmd=rcmboot
 else
-    if [ -z "$sdcard" -a $no_flash -eq 0 -a "$spi_only" != "yes" ]; then
+    if [ -z "$sdcard" -a $external_device -eq 0 -a $no_flash -eq 0 -a "$spi_only" != "yes" ]; then
         rm -f "$appfile"
         echo "Creating sparseimage ${appfile}..."
         $here/mksparse -b ${blocksize} --fillpattern=0 "$imgfile" "$appfile" || exit 1
@@ -414,7 +480,11 @@ else
 fi
 
 temp_user_dir=
-if [ -n "$keyfile" ]; then
+want_signing=0
+if [ -n "$keyfile" ] || [ $rcm_boot -eq 1 ] || [ $no_flash -eq 1 -a $to_sign -eq 1 ]; then
+    want_signing=1
+fi
+if [ $have_odmsign_func -eq 1 -a $want_signing -eq 1 ]; then
     if [ -n "$sbk_keyfile" ]; then
         if [ -z "$user_keyfile" ]; then
             rm -f "null_user_key.txt"
@@ -462,18 +532,22 @@ if [ -n "$keyfile" ]; then
     bpfdtbfilename="$BPFDTB_FILE"
     localbootfile="$kernfile"
     BINSARGS="--bins \"$BINSARGS\""
-    flashername=uefi_jetson_with_dtb.bin
+    flashername="uefi_jetson_with_dtb.bin"
+    UEFIBL="uefi_jetson_with_dtb.bin"
+    mb1filename="mb1_t234_prod.bin"
+    pscbl1filename="psc_bl1_t234_prod.bin"
+    tbcfilename="uefi_jetson.bin"
+    custinfofilename="custinfo_out.bin"
+    BL_DIR="."
     BCT="--sdram_config"
     bctfilename=`echo $sdramcfg_files | cut -d, -f1`
     bctfile1name=`echo $sdramcfg_files | cut -d, -f2`
     SOSARGS="--applet mb1_t234_prod.bin "
     NV_ARGS=" "
-    BCTARGS="$bctargs"
+    BCTARGS="$bctargs --boot_chain A --bct_backup --secondary_gpt_backup"
     rootfs_ab=0
-    rcm_boot=0
-    external_device=0
     . "$here/odmsign.func"
-    (odmsign_ext) || exit 1
+    (odmsign_ext_sign_and_flash) || exit 1
     if [ $bup_blob -eq 0 -a $no_flash -ne 0 ]; then
         if [ -f flashcmd.txt ]; then
             chmod +x flashcmd.txt
@@ -494,7 +568,7 @@ if [ -n "$keyfile" ]; then
 fi
 
 flashcmd="python3 $flashappname ${inst_args} --chip 0x23 --bl uefi_jetson_with_dtb.bin \
-          --sdram_config $sdramcfg_files
+          --sdram_config $sdramcfg_files \
           --odmdata $odmdata \
           --applet mb1_t234_prod.bin \
           --cmd \"$tfcmd\" $skipuid \
@@ -502,7 +576,7 @@ flashcmd="python3 $flashappname ${inst_args} --chip 0x23 --bl uefi_jetson_with_d
           --boot_chain A \
           --bct_backup \
           --secondary_gpt_backup \
-          $bctargs \
+          $bctargs $extdevargs \
           --bins \"$BINSARGS\""
 
 if [ $bup_blob -ne 0 ]; then
@@ -521,9 +595,28 @@ if [ $bup_blob -ne 0 ]; then
         exit 1
     fi
     l4t_bup_gen "$flashcmd" "$spec" "$fuselevel" t186ref "$keyfile" "$sbk_keyfile" 0x23 || exit 1
+    exit 0
+fi
+
+if [ $to_sign -ne 0 ]; then
+    eval $flashcmd < /dev/null || exit 1
+    exit 0
+fi
+
+if [ $no_flash -ne 0 ]; then
+    echo "$flashcmd" | sed -e 's,--skipuid,,g' > flashcmd.txt
+    chmod +x flashcmd.txt
+    rm -f APPFILE APPFILE_b DATAFILE null_user_key.txt
 else
     eval $flashcmd < /dev/null || exit 1
-    if [ -n "$sdcard" ]; then
+    if [ -n "$sdcard" -o $external_device -eq 1 ]; then
+	if [ $external_device -eq 1 ]; then
+	    if [ -z "$serial_number" ]; then
+		echo "ERR: missing serial number for initrd-flashing external device" >&2
+		exit 1
+	    fi
+	    make_sdcard_args="$make_sdcard_args --serial-number $serial_number"
+	fi
         if [ -n "$pre_sdcard_sed" ]; then
             rm -f signed/flash.xml.tmp.in
             mv signed/flash.xml.tmp signed/flash.xml.tmp.in
